requirements/base.txt
django==1.11
django-environ==0.4.3

# Time zones support
pytz==2017.2
